Create vpn.ts
