+++ client/src/www/app/main.cordova.ts
@@ -27,7 +27,7 @@ import {AbstractClipboard} from './clipboard';
 import {EnvironmentVariables} from './environment';
 import {main} from './main';
 import {OutlineServerRepository} from './outline_server_repository';
-import {CordovaTunnel} from './outline_server_repository/server.cordova';
+import {CordovaTunnel} from './outline_server_repository/vpn.cordova';
 import {OutlinePlatform} from './platform';
 import {OUTLINE_PLUGIN_NAME, pluginExec} from './plugin.cordova';
 import {AbstractUpdater} from './updater';
