+++ client/src/www/app/main.ts
@@ -20,7 +20,7 @@ import {makeConfig, SIP002_URI} from 'ShadowsocksConfig';
 import {App} from './app';
 import {onceEnvVars} from './environment';
 import {OutlineServerRepository} from './outline_server_repository';
-import {FAKE_BROKEN_HOSTNAME, FAKE_UNREACHABLE_HOSTNAME, FakeTunnel} from './outline_server_repository/server.fake';
+import {FAKE_BROKEN_HOSTNAME, FAKE_UNREACHABLE_HOSTNAME, FakeTunnel} from './outline_server_repository/vpn.fake.js';
 import {OutlinePlatform} from './platform';
 import {Settings} from './settings';
 import {EventQueue} from '../model/events';
