+++ client/src/www/app/outline_server_repository/server.ts
@@ -19,6 +19,7 @@ import * as errors from '../../model/errors';
 import * as events from '../../model/events';
 import {PlatformError} from '../../model/platform_error';
 import {Server, ServerType} from '../../model/server';
+import { ShadowsocksSessionConfig, PlatformTunnel, TunnelStatus } from './vpn';
 
 
 // PLEASE DON'T use this class outside of this `outline_server_repository` folder!
@@ -147,40 +148,4 @@ export class OutlineServer implements Server {
 }
 
 
-export interface ShadowsocksSessionConfig {
-  host?: string;
-  port?: number;
-  password?: string;
-  method?: string;
-  prefix?: string;
-}
 
-export const enum TunnelStatus {
-  CONNECTED,
-  DISCONNECTED,
-  RECONNECTING,
-}
-
-export type TunnelFactory = (id: string) => PlatformTunnel;
-
-// Represents a VPN tunnel to a Shadowsocks proxy server. Implementations provide native tunneling
-// functionality through cordova.plugins.oultine.Tunnel and ElectronOutlineTunnel.
-export interface PlatformTunnel {
-  // Unique instance identifier.
-  readonly id: string;
-
-  // Connects a VPN, routing all device traffic to a Shadowsocks server as dictated by `config`.
-  // If there is another running instance, broadcasts a disconnect event and stops the active
-  // tunnel. In such case, restarts tunneling while preserving the VPN.
-  // Throws OutlinePluginError.
-  start(name: string, config: ShadowsocksSessionConfig): Promise<void>;
-
-  // Stops the tunnel and VPN service.
-  stop(): Promise<void>;
-
-  // Returns whether the tunnel instance is active.
-  isRunning(): Promise<boolean>;
-
-  // Sets a listener, to be called when the tunnel status changes.
-  onStatusChange(listener: (status: TunnelStatus) => void): void;
-}
