+++ client/src/www/app/outline_server_repository/vpn.cordova.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {ShadowsocksSessionConfig, PlatformTunnel, TunnelStatus} from './server';
+import { ShadowsocksSessionConfig, PlatformTunnel, TunnelStatus } from './vpn';
 import * as errors from '../../model/errors';
 import {OUTLINE_PLUGIN_NAME, pluginExecWithErrorCode} from '../plugin.cordova';
