+++ client/src/www/app/outline_server_repository/vpn.fake.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {PlatformTunnel, TunnelStatus, ShadowsocksSessionConfig} from './server';
+import { PlatformTunnel, TunnelStatus, ShadowsocksSessionConfig } from './vpn';
 import * as errors from '../../model/errors';
 
 export const FAKE_BROKEN_HOSTNAME = '192.0.2.1';
